refactor: move user game state from user package to utils
